broke out functions into functions.js. cleaned up app.js. hit api limit for the day.
